Génération du bordereau sur un seul fichier
